Move EmptyComponents, which is used by common/test and also by other packages
